feat: optmize asset drawer state initialization
