+++ frontend/pages/home/components/ICRC/asset/AssetDrawerMutate.tsx
@@ -1,16 +1,3 @@
-import PlusIcon from "@assets/svg/files/plus-icon.svg";
-// svgs
-import {
-  AssetMutationAction,
-  AssetMutationResult,
-  setAccordionAssetIdx,
-  setAssetMutation,
-  setAssetMutationAction,
-  setAssetMutationResult,
-  setSelectedAsset,
-} from "@redux/assets/AssetReducer";
-import { useAppDispatch, useAppSelector } from "@redux/Store";
-import { useMemo } from "react";
 import { BasicDrawer } from "@components/drawer";
 import CloseAssetDrawer from "./CloseAssetDrawer";
 import AddAssetAutomatic from "./AddAssetAutomatic";
@@ -21,69 +8,67 @@ import { Asset } from "@redux/models/AccountModels";
 import { getAssetDetails } from "@pages/home/helpers/icrc";
 import { db } from "@/database/db";
 import { useTranslation } from "react-i18next";
+import { useMemo } from "react";
+import {
+  AssetMutationAction,
+  AssetMutationResult,
+  setAccordionAssetIdx,
+  setAssetMutation,
+  setAssetMutationAction,
+  setAssetMutationResult,
+  setSelectedAsset,
+} from "@redux/assets/AssetReducer";
+import { useAppDispatch, useAppSelector } from "@redux/Store";
 
-export default function AssetDrawerMutate() {
+export default function AssetDrawerMutate({ isDrawerOpen }: { isDrawerOpen: boolean }) {
   const dispatch = useAppDispatch();
   const { t } = useTranslation();
   const { assets } = useAppSelector((state) => state.asset);
-  const { accordionIndex } = useAppSelector((state) => state.asset.helper);
   const { assetMutated, assetAction } = useAppSelector((state) => state.asset.mutation);
+  const { accordionIndex } = useAppSelector((state) => state.asset.helper);
 
   const { newAsset, setNewAsset, setValidToken, setErrIndex, setErrToken, errIndex, errToken, validToken } =
     useAssetMutate();
 
   const isManual = useMemo(() => assetAction === AssetMutationAction.ADD_MANUAL, [assetAction]);
   const isUpdate = useMemo(() => assetAction === AssetMutationAction.UPDATE, [assetAction]);
-  const isDrawerOpen = useMemo(
-    () => assetAction !== AssetMutationAction.NONE && assetAction !== AssetMutationAction.DELETE,
-    [assetAction],
-  );
 
   return (
-    <>
-      <div className={containerStyles} onClick={onAddAsset}>
-        <img src={PlusIcon} alt="plus-icon" />
+    <BasicDrawer isDrawerOpen={isDrawerOpen}>
+      <div className={formContainerStyles}>
+        <CloseAssetDrawer onClose={onClose} isEdit={assetMutated ? true : false} />
+
+        {isManual || isUpdate ? (
+          <AddAssetManual
+            errToken={errToken}
+            setErrToken={setErrToken}
+            errIndex={errIndex}
+            setErrIndex={setErrIndex}
+            validToken={validToken}
+            setValidToken={setValidToken}
+            newAsset={newAsset}
+            setNewAsset={setNewAsset}
+            asset={assetMutated}
+            addAssetToData={addAssetToData}
+          ></AddAssetManual>
+        ) : (
+          <AddAssetAutomatic
+            setNewAsset={setNewAsset}
+            newAsset={newAsset}
+            addAssetToData={addAssetToData}
+            setValidToken={setValidToken}
+            setErrToken={setErrToken}
+            errToken={errToken}
+          ></AddAssetAutomatic>
+        )}
+
+        <DialogAssetConfirmation newAsset={newAsset} />
       </div>
-      {isDrawerOpen && (
-        <BasicDrawer isDrawerOpen={isDrawerOpen}>
-          <div className={formContainerStyles}>
-            <CloseAssetDrawer onClose={onClose} isEdit={assetMutated ? true : false} />
-
-
-            {isManual || isUpdate ? (
-              <AddAssetManual
-                errToken={errToken}
-                setErrToken={setErrToken}
-                errIndex={errIndex}
-                setErrIndex={setErrIndex}
-                validToken={validToken}
-                setValidToken={setValidToken}
-                newAsset={newAsset}
-                setNewAsset={setNewAsset}
-                asset={assetMutated}
-                addAssetToData={addAssetToData}
-              ></AddAssetManual>
-            ) : (
-              <AddAssetAutomatic
-                setNewAsset={setNewAsset}
-                newAsset={newAsset}
-                addAssetToData={addAssetToData}
-                setValidToken={setValidToken}
-                setErrToken={setErrToken}
-                errToken={errToken}
-              ></AddAssetAutomatic>
-            )}
-
-            <DialogAssetConfirmation newAsset={newAsset} />
-
-          </div>
-        </BasicDrawer>
-      )}
-    </>
+    </BasicDrawer>
   );
 
-  function onAddAsset() {
-    dispatch(setAssetMutationAction(AssetMutationAction.ADD_AUTOMATIC));
+  function isAssetAdded(address: string) {
+    return assets.find((asst: Asset) => asst.address === address) ? true : false;
   }
 
   function onClose() {
@@ -94,10 +79,6 @@ export default function AssetDrawerMutate() {
     dispatch(setAssetMutation(undefined));
     dispatch(setAssetMutationAction(AssetMutationAction.NONE));
   }
-
-  function isAssetAdded(address: string) {
-    return assets.find((asst: Asset) => asst.address === address) ? true : false;
-  };
 
   async function addAssetToData() {
     if (isAssetAdded(newAsset.address)) {
@@ -125,7 +106,6 @@ export default function AssetDrawerMutate() {
         dispatch(setAssetMutationResult(AssetMutationResult.ADDED));
         dispatch(setSelectedAsset(assetToSave));
         dispatch(setAccordionAssetIdx([assetToSave.symbol]));
-
       } catch (error) {
         console.error("Error adding asset", error);
         dispatch(setAssetMutationResult(AssetMutationResult.FAILED));
@@ -139,8 +119,6 @@ export default function AssetDrawerMutate() {
     }
   }
 }
-
-const containerStyles = "flex flex-row items-center justify-center w-8 h-8 rounded-md cursor-pointer bg-SelectRowColor";
 
 const formContainerStyles =
   "px-8 mt-4 overflow-y-auto text-left text-PrimaryTextColorLight dark:text-PrimaryTextColor text-md";
